[Feature] Add ForumatePlus (#60)
