+++ Forumate.xcodeproj/xcshareddata/xcschemes/Forumate.xcscheme
@@ -62,6 +62,9 @@
             ReferencedContainer = "container:Forumate.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
+      <StoreKitConfigurationFileReference
+         identifier = "../../Forumate/Store/Forumate.storekit">
+      </StoreKitConfigurationFileReference>
    </LaunchAction>
    <ProfileAction
       buildConfiguration = "Release"
