+++ Forumate/Other/Localizable.xcstrings
@@ -77,6 +77,16 @@
         }
       }
     },
+    "1" : {
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : ""
+          }
+        }
+      }
+    },
     "Acknowledgement" : {
       "localizations" : {
         "zh-Hans" : {
@@ -113,6 +123,16 @@
           "stringUnit" : {
             "state" : "translated",
             "value" : "应用图标"
+          }
+        }
+      }
+    },
+    "App Icon Replacement" : {
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "应用图标更换"
           }
         }
       }
@@ -173,6 +193,16 @@
           "stringUnit" : {
             "state" : "translated",
             "value" : "取消"
+          }
+        }
+      }
+    },
+    "Close" : {
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "关闭"
           }
         }
       }
@@ -287,6 +317,16 @@
         }
       }
     },
+    "Development Support Subscription" : {
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "开发支持套餐"
+          }
+        }
+      }
+    },
     "Discuss at Discourse Forum Topic" : {
       "localizations" : {
         "zh-Hans" : {
@@ -313,6 +353,16 @@
           "stringUnit" : {
             "state" : "translated",
             "value" : "例子： forums.example.com"
+          }
+        }
+      }
+    },
+    "Exclusive features" : {
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "专属功能"
           }
         }
       }
@@ -357,6 +407,16 @@
         }
       }
     },
+    "Forumate's most feature is free and its source code is avaiable on GitHub." : {
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "Forumate 的大部分功能是免费的，其源代码可在 GitHub 上查看。"
+          }
+        }
+      }
+    },
     "Forumate+" : {
       "localizations" : {
         "zh-Hans" : {
@@ -373,6 +433,16 @@
           "stringUnit" : {
             "state" : "translated",
             "value" : "通用"
+          }
+        }
+      }
+    },
+    "Hello, World!" : {
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : ""
           }
         }
       }
@@ -413,6 +483,26 @@
           "stringUnit" : {
             "state" : "translated",
             "value" : "加入 Forumate 的 Discord 社区"
+          }
+        }
+      }
+    },
+    "Learn more about Forumate+" : {
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "了解更多关于 Forumate+ 的信息"
+          }
+        }
+      }
+    },
+    "Lifetime Plan" : {
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "永久套餐"
           }
         }
       }
@@ -543,6 +633,16 @@
           "stringUnit" : {
             "state" : "translated",
             "value" : "好的"
+          }
+        }
+      }
+    },
+    "Only avaiable on supported platforms" : {
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "仅在受支持的平台上可用"
           }
         }
       }
@@ -687,6 +787,16 @@
         }
       }
     },
+    "Subscription Plan" : {
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "订阅套餐"
+          }
+        }
+      }
+    },
     "Successfully saved the icon to the Downloads folder." : {
       "localizations" : {
         "zh-Hans" : {
@@ -807,12 +917,12 @@
         }
       }
     },
-    "Update App Icon in App is not supported on macOS. You can click to save the icon and manully update it via Finder." : {
+    "Update App Icon is not supported on macOS. But you can click to save the icon and manully update it via Finder." : {
       "localizations" : {
         "zh-Hans" : {
           "stringUnit" : {
             "state" : "translated",
-            "value" : "macOS 不支持在应用中更新图标。您可以单击以保存图标，并通过Finder进行手动更新。"
+            "value" : "macOS 不支持更新图标。但您可以单击以保存图标，并通过Finder进行手动更新。"
           }
         }
       }
@@ -843,6 +953,67 @@
           "stringUnit" : {
             "state" : "translated",
             "value" : "欢迎使用 Forumate"
+          }
+        }
+      }
+    },
+    "You can support this project by becoming a Forumate+ Member via **Subscription Plan** or **Lifetime Plan**" : {
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "您可以通过**订阅套餐**或**永久套餐**成为 Forumate+ 会员来支持这个项目"
+          }
+        }
+      }
+    },
+    "You do not have any active Forumate+ subscription" : {
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "您没有任何活跃的 Forumate+ 订阅"
+          }
+        }
+      }
+    },
+    "You have a lifetime Forumate+ membership." : {
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "您拥有永久 Forumate+ 会员身份。"
+          }
+        }
+      }
+    },
+    "You have an active Forumate+ subscription. Thank you for your support." : {
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "您有一个活跃的 Forumate+ 订阅。感谢您的支持。"
+          }
+        }
+      }
+    },
+    "You need an active Forumate+ membership to unlock this feature." : {
+      "comment" : "Plus only feature Description",
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "您需要一个活跃的 Forumate+ 会员身份才能解锁此功能。"
+          }
+        }
+      }
+    },
+    "Your Forumate+ subscription is expired" : {
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "您的 Forumate+ 订阅已过期"
           }
         }
       }
