+++ Forumate/Store/Forumate.storekit
@@ -5,18 +5,18 @@
   ],
   "products" : [
     {
-      "displayPrice" : "9.99",
+      "displayPrice" : "14.99",
       "familyShareable" : true,
       "internalID" : "6450400378",
       "localizations" : [
         {
-          "description" : "Pro Version of Forumate",
-          "displayName" : "Forumate Pro",
+          "description" : "Lifetime membership for Forumate+",
+          "displayName" : "Lifetime Forumate+",
           "locale" : "en_US"
         }
       ],
       "productID" : "nonconsumable.pro_version",
-      "referenceName" : "ProVersion",
+      "referenceName" : "Lifetime",
       "type" : "NonConsumable"
     }
   ],
@@ -24,22 +24,31 @@
     "_applicationInternalID" : "6449866904",
     "_developerTeamID" : "VB7MJ8R223",
     "_failTransactionsEnabled" : false,
-    "_lastSynchronizedDate" : 708806917.31874597,
+    "_lastSynchronizedDate" : 715940801.67057896,
     "_locale" : "en_US",
     "_storefront" : "USA",
     "_storeKitErrors" : [
       {
-        "current" : null,
+        "current" : {
+          "index" : 2,
+          "type" : "generic"
+        },
         "enabled" : false,
         "name" : "Load Products"
       },
       {
-        "current" : null,
+        "current" : {
+          "index" : 1,
+          "type" : "purchase"
+        },
         "enabled" : false,
         "name" : "Purchase"
       },
       {
-        "current" : null,
+        "current" : {
+          "index" : 0,
+          "type" : "verification"
+        },
         "enabled" : false,
         "name" : "Verification"
       },
@@ -49,12 +58,18 @@
         "name" : "App Store Sync"
       },
       {
-        "current" : null,
+        "current" : {
+          "index" : 2,
+          "type" : "generic"
+        },
         "enabled" : false,
         "name" : "Subscription Status"
       },
       {
-        "current" : null,
+        "current" : {
+          "index" : 3,
+          "type" : "generic"
+        },
         "enabled" : false,
         "name" : "App Transaction"
       },
@@ -64,7 +79,10 @@
         "name" : "Manage Subscriptions Sheet"
       },
       {
-        "current" : null,
+        "current" : {
+          "index" : 0,
+          "type" : "refund request"
+        },
         "enabled" : false,
         "name" : "Refund Request Sheet"
       },
@@ -94,17 +112,52 @@
           "familyShareable" : true,
           "groupNumber" : 1,
           "internalID" : "6450400527",
-          "introductoryOffer" : null,
+          "introductoryOffer" : {
+            "internalID" : "4FA5ADE5",
+            "numberOfPeriods" : 1,
+            "paymentMode" : "free",
+            "subscriptionPeriod" : "P1W"
+          },
           "localizations" : [
             {
-              "description" : "Show your support to the app development",
-              "displayName" : "Development Support",
+              "description" : "Support development by a monthly subscription",
+              "displayName" : "Forumate+ Month Plan",
               "locale" : "en_US"
             }
           ],
           "productID" : "subscribe.pro.monthly",
           "recurringSubscriptionPeriod" : "P1M",
           "referenceName" : "monthly",
+          "subscriptionGroupID" : "21353237",
+          "type" : "RecurringSubscription"
+        },
+        {
+          "adHocOffers" : [
+
+          ],
+          "codeOffers" : [
+
+          ],
+          "displayPrice" : "9.99",
+          "familyShareable" : true,
+          "groupNumber" : 2,
+          "internalID" : "6464517034",
+          "introductoryOffer" : {
+            "internalID" : "C234686E",
+            "numberOfPeriods" : 1,
+            "paymentMode" : "free",
+            "subscriptionPeriod" : "P1M"
+          },
+          "localizations" : [
+            {
+              "description" : "Support development by a yearly subscription",
+              "displayName" : "Forumate+ Year Plan",
+              "locale" : "en_US"
+            }
+          ],
+          "productID" : "subscribe.pro.yearly",
+          "recurringSubscriptionPeriod" : "P1Y",
+          "referenceName" : "yearly",
           "subscriptionGroupID" : "21353237",
           "type" : "RecurringSubscription"
         }
