+++ Forumate/View/Tab/Settings/Sections/PrivacyPolicy/PrivacyPolicySection.swift
